# finish/src/main/webapp/META-INF/openapi.yaml
openapi: 3.0.0
info:
  title: Deployed APIs
  version: 1.0.0
servers:
- url: http://localhost:9080
paths:
  /inventory/systems/{hostname}:
    get:
      summary: Get JVM system properties for particular host
      description: Retrieves and returns the JVM system properties from the system
        service running on the particular host.
      operationId: getPropertiesForHost
      parameters:
      - name: hostname
        in: path
        description: The host for whom to retrieve the JVM system properties for.
        required: true
        schema:
          type: string
        example: foo
      responses:
        404:
          description: Missing description - to be filtered.
          content:
            none: {}
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Properties'
  /inventory/systems:
    get:
      summary: List inventory contents.
      description: Returns the currently stored host:properties pairs in the inventory.
      operationId: listContents
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryList'
  /inventory/properties:
    get:
      summary: Get JVM system properties.
      description: Returns the JVM system properties for the host running this service.
      operationId: getProperties
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Properties'
components:
  schemas:
    InventoryList:
      required:
      - systems
      type: object
      properties:
        systems:
          type: array
          items:
            $ref: '#/components/schemas/System'
        total:
          type: integer
      description: POJO that represents the inventory contents.
    Properties:
      type: object
      additionalProperties:
        type: string
    System:
      required:
      - hostname
      - properties
      type: object
      properties:
        hostname:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
      description: POJO that represents a single inventory entry.